refactor: move Plex domain retrieval to utils module for better code organization
- Refactored the Plex domain retrieval logic into a new utils.js module, enhancing code reusability and maintainability.
- Updated client.js and thumb.js to utilize the new getPlexDomain function from utils, streamlining the codebase.
- Improved error handling during the domain retrieval process.
